Resolving merge conflicts
